contracts-bedrock: fix custom gas token withdrawals to L1
The previous implementation of the `FeeVault` used the `StandardBridge`
to withdraw `ether`. Under custom gas token, this meant that withdrawals
to L1 didn't work and withdrawals to L2 were the only supported path.
This updates the `L2ToL1Messenger` to withdraw to L1 via the `L2ToL1Messenger`.
This fixes the ability to withdraw to L1 while custom gas token is
enabled while also saving some gas on both L1 and L2 by removing
extra call frames. The reason why this was originally implemented using
the `StandardBridge` was to reduce additional changes during the
development of bedrock so that more effort could be spent thinking
about the new protocol design, as the legacy implementation used
the `StandardBridge` for doing the withdrawal.

Having the `StandardBridge` used for `ether` deposits and withdrawals
is unncessarily expensive and a bit of a leaky abstraction.

Gas savings are scalability improvements, both on L1 and on L2.

Also uses a new variant of `SafeCall` that passes through all gas
in its `transfer`. The implementation ensures to assert on the return
value of the call to hold the invariant true that the `totalProcessed`
will always be consistent. If the call reverts, then its possible for
the `ether` to not be transferred and `totalProcessed` to be incremented.

Also bump the semver of the contracts to a beta version. They will
be bumped out of beta at a release of the contracts.

The tests are also improved for both the `SafeCall` library as well
as the `FeeVault`.